+++ compile-davinci-helper.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/bash
 clear
-version="2.4.5"
+version="2.4.6"
 
 if [ -d "/home/$USER/build_davinci_helper" ]; then
     rm -rf "/home/$USER/build_davinci_helper"
