+++ create_archive.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/bash
 clear
-version="2.4.5"
+version="2.4.6"
 
 mv davinci-helper davinci-helper-${version}
 tar -cvzf davinci-helper-${version}.tar.gz davinci-helper-${version}
