+++ davinci-helper/davinci_helper/functions/gui/loading_gui.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2024 Lorenzo Maiuri
+# Copyright 2025 Lorenzo Maiuri
 # Pubblicato sotto licenza GPL-3.0
 # Published under GPL-3.0 license
 # GitHub : https://github.com/H3rz3n/davinci-helper
