+++ davinci-helper/davinci_helper/functions/logic/install_ffmpeg.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2024 Lorenzo Maiuri
+# Copyright 2025 Lorenzo Maiuri
 # Pubblicato sotto licenza GPL-3.0
 # Published under GPL-3.0 license
 #   
@@ -17,6 +17,10 @@
 
 # STANDARD MODULES IMPORT
 import sys, os, subprocess, threading, gettext, locale
+
+# NOT STANDARD MODULES IMPORT
+from .utility import utility
+from .function_4 import function_4
 
 #-----------------------------------------------------------------------------------------------------
 
@@ -45,6 +49,7 @@ _ = gettext.gettext
 
 
 
+
 # FUNCTION THAT WILL CHECK IF FFMPEG IS ALREADY INSTALLED ON THE SYSTEM
 def check_ffmpeg_presence ():
 
@@ -70,56 +75,6 @@ def check_ffmpeg_presence ():
         return "None"
 
     #-----------------------------------------------------------------------------------------------------
-
-
-
-
-
-# FUNCTION THAT ADDS THE RPM FUSION REPOSITORY
-def add_repository():
-
-    #-----------------------------------------------------------------------------------------------------
-
-    # ACQUIRING IF RPM FUSION REPO IS ALREADY INSTALLED
-    rpm_fusion_repo_check = subprocess.run("dnf repolist", shell=True, capture_output=True, text=True )
-
-    # CHECKING IF RPM FUSION REPO IS ALREADY INSTALLED
-    if rpm_fusion_repo_check.stdout.find("rpmfusion-free") != -1 and rpm_fusion_repo_check.stdout.find("rpmfusion-nonfree"):
-
-        # PRINTING THE MESSAGE
-        print(_("The RPM Fusion repository had already been added to the system, there was no need to add it."))
-        print("")
-
-    else :
-    
-        # ADDING THE RPM FUSION REPOSITORY
-        adding_repo = subprocess.run("dnf install -y https://mirrors.rpmfusion.org/free/fedora/rpmfusion-free-release-$(rpm -E %fedora).noarch.rpm https://mirrors.rpmfusion.org/nonfree/fedora/rpmfusion-nonfree-release-$(rpm -E %fedora).noarch.rpm", shell=True, capture_output=True, text=True )
-
-        # CHECKING IF THERE WERE ERRORS
-        if adding_repo.returncode != 0 :
-
-            # PRINTING THE ERROR MESSAGE
-            print(_("DEBUG : It was impossible to add the RPM Fusion Free and Non-Free repository."))
-            print(_("Check your network connection and try again or add it by yourself."))
-            print("")
-            print(adding_repo.stdout)
-            print("")
-            print(_("Please open an issue report and paste this error code on the project GitHub page :"))
-            print("")
-            print("https://github.com/H3rz3n/davinci-helper/issues")
-            print("")
-            exit(2)
-
-        else :
-
-            # PRINTING THE SUCCESSFUL STATE
-            print(_("The RPM Fusion Free and Non-Free repository have been successfully added."))
-            print("")
-
-    #-----------------------------------------------------------------------------------------------------
-
-
-
 
 
 
@@ -152,8 +107,6 @@ def install_ffmpeg ():
         print("")
 
     #-----------------------------------------------------------------------------------------------------
-
-
 
 
 
@@ -194,7 +147,7 @@ def swap_ffmpeg ():
 
 
 # FUNCTION THAT WILL CHECK ALL THE CODECS ARE ALREADY INSTALLED ON THE SYSTEM
-def check_multimedia_codecs_presence ():
+def check_multimedia_codecs_presence (os_version):
 
     #-----------------------------------------------------------------------------------------------------
 
@@ -243,10 +196,6 @@ def check_multimedia_codecs_presence ():
     # SEARCHING
     if check_codecs.stdout.find('gstreamer1-plugins-ugly-free') == -1 :
         lib_to_install = lib_to_install + " gstreamer1-plugins-ugly-free"
-    
-    # SEARCHING
-    if check_codecs.stdout.find('libva-intel-media-driver') == -1 :
-        lib_to_install = lib_to_install + " libva-intel-media-driver"
 
     # SEARCHING
     if check_codecs.stdout.find('pipewire-alsa') == -1 :
@@ -437,6 +386,11 @@ is_installed_ffmpeg = check_ffmpeg_presence()
 
 #-----------------------------------------------------------------------------------------------------
 
+# ACQUIRING WHICH VERSION OF FEDORA IS IN USE
+os_version = check_fedora_version()
+
+#-----------------------------------------------------------------------------------------------------
+
 # CHECKING IF FFMPEG FROM RPM FUSION IS INSTALLED
 if is_installed_ffmpeg == "Lite" :
 
@@ -475,126 +429,10 @@ else :
 
 
 
-
-# FUNCTION THAT CHECK WHICH VERSION OF FEDORA IS INSTALLED
-def check_fedora_version ():
-
-    #-----------------------------------------------------------------------------------------------------
-    
-    # READING WHICH VERSION OF FEDORA IS INSTALLED
-    os_info = subprocess.run("hostnamectl", shell=True, capture_output=True, text=True)
-
-    # PRINTING IN THE TERMINAL THE RESULT DEPENDING ON WHETHER THERE ARE ERRORS OR NOT
-    if os_info.returncode == 0: 
-
-        # CHECKING WHIC VERSION OF FEDORA IS USED
-        if os_info.stdout.find("Fedora Linux 40") != -1 :
-        
-            # SETTING THE FOUND OS VERSION
-            os_version = "Fedora Linux 40"
-
-            # PRINT THE FEDORA VERSION
-            print(_("You are using a supported OS version : {os_version_placeholder}").format(os_version_placeholder = os_version))
-        
-        elif os_info.stdout.find("Fedora Linux 41") != -1 :
-            
-            # SETTING THE FOUND OS VERSION
-            os_version = "Fedora Linux 41"
-
-            # PRINT THE FEDORA VERSION
-            print(_("You are using a supported OS version : {os_version_placeholder}").format(os_version_placeholder = os_version))
-
-        elif os_info.stdout.find("Fedora Linux 42") != -1 :
-
-            # SETTING THE FOUND OS VERSION
-            os_version = "Fedora Linux 42"
-
-            # PRINT THE FEDORA VERSION
-            print(_("You are using a supported OS version : {os_version_placeholder}").format(os_version_placeholder = os_version))
-
-        elif ((os_info.stdout).lower()).find("rawhide") != -1 :
-
-            # SETTING THE FOUND OS VERSION
-            os_version = "Fedora Linux Rawhide"
-
-            # PRINT THE FEDORA VERSION
-            print(_("You are using a supported OS version : {os_version_placeholder}").format(os_version_placeholder = os_version))
-        
-        elif os_info.stdout.find("Nobara Linux 40") != -1 :
-
-            # SETTING THE FOUND OS VERSION
-            os_version = "Nobara Linux 40"
-
-            # PRINT THE FEDORA VERSION
-            print(_("You are using a supported OS version : {os_version_placeholder}").format(os_version_placeholder = os_version))
-
-        elif os_info.stdout.find("Nobara Linux 41") != -1 :
-
-            # SETTING THE FOUND OS VERSION
-            os_version = "Nobara Linux 41"
-
-            # PRINT THE FEDORA VERSION
-            print(_("You are using a supported OS version : {os_version_placeholder}").format(os_version_placeholder = os_version))
-
-        elif os_info.stdout.find("Nobara Linux 42") != -1 :
-
-            # SETTING THE FOUND OS VERSION
-            os_version = "Nobara Linux 42"
-
-            # PRINT THE FEDORA VERSION
-            print(_("You are using a supported OS version : {os_version_placeholder}").format(os_version_placeholder = os_version))
-
-        elif os_info.stdout.find("Ultramarine Linux 40") != -1 :
-
-            # SETTING THE FOUND OS VERSION
-            os_version = "Ultramarine Linux 40"
-
-            # PRINT THE FEDORA VERSION
-            print(_("You are using a supported OS version : {os_version_placeholder}").format(os_version_placeholder = os_version))
-
-        elif os_info.stdout.find("Ultramarine Linux 41") != -1 :
-
-            # SETTING THE FOUND OS VERSION
-            os_version = "Ultramarine Linux 41"
-
-            # PRINT THE FEDORA VERSION
-            print(_("You are using a supported OS version : {os_version_placeholder}").format(os_version_placeholder = os_version))
-
-        elif os_info.stdout.find("Ultramarine Linux 42") != -1 :
-
-            # SETTING THE FOUND OS VERSION
-            os_version = "Ultramarine Linux 42"
-
-            # PRINT THE FEDORA VERSION
-            print(_("You are using a supported OS version : {os_version_placeholder}").format(os_version_placeholder = os_version))
-
-
-        
-
-        # RETURNS VALUE TO THE SCRIPT
-        return os_version
-        
-    else:
-        print(_("DEBUG : There was an error reading what OS is installed :"))
-        print("")
-        print(os_info.stderr)
-        print("")
-        print(_("Please open an issue report and paste this error code on the project GitHub page :"))
-        print("")
-        print("https://github.com/H3rz3n/davinci-helper/issues")
-        print("")
-        exit(1)
-
-    #-----------------------------------------------------------------------------------------------------
-
-
-
-
-
 #-----------------------------------------------------------------------------------------------------
 
 # ACQUIRING IF THE CODECS FROM RPM FUSION ARE INSTALLED
-is_installed_multimedia_codecs, lib_install_list = check_multimedia_codecs_presence()
+is_installed_multimedia_codecs, lib_install_list = check_multimedia_codecs_presence(os_version)
 
 #-----------------------------------------------------------------------------------------------------
 
@@ -609,9 +447,6 @@ if is_installed_multimedia_codecs == "None" :
     # INSTALLING FFMPEG
     install_codecs(lib_install_list)
 
-    # ACQUIRING THE OS VERSION
-    os_version = check_fedora_version()
-
     # CHECKING WHICH VERSION OF THE OS IS IN USE
     if (os_version.find("40") != -1 or os_version.find("41") != -1) :
 
@@ -621,7 +456,7 @@ if is_installed_multimedia_codecs == "None" :
     else :
 
         # UPDATE MULTIMEDIA GROUP
-        update_multimedia__42()
+        update_multimedia_42()
         
     exit(0)
